requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
celery>=5.0,<6.0
redis>=3.5,<4.0
gunicorn>=20.1,<21.0
psycopg2-binary>=2.8,<3.0
